refactor: split large components into modular directories
- gallery.tsx: 1097 → 617 lines (44% reduction)
  - Extracted: types, utils, translations, thumbnail-item, location-map, sidebar-panel, mobile-drawer

- stats-dashboard.tsx: 905 → 617 lines (32% reduction)
  - Extracted: types, animated-counter, code-frequency-heatmap

- photo-viewer.tsx: 990 → 954 lines
  - Extracted: types, utils

- moment-detail.tsx: 624 → 588 lines
  - Extracted: types, translations

Total: 3616 → 2776 lines (23% reduction, 840 lines saved)
