+++ src/components/photo-viewer/index.ts
@@ -1,0 +1,4 @@
+// Photo viewer component barrel export
+
+export { PhotoViewer } from "./photo-viewer-main";
+export type { PhotoViewerProps, OriginalLoadState } from "./types";
