+++ src/components/zhi/gallery/index.ts
@@ -1,0 +1,6 @@
+// Gallery component barrel export
+// This is the main entry point for the gallery component
+
+export { ZhiGallery as default } from "./gallery-main";
+export { ZhiGallery } from "./gallery-main";
+export type { ZhiGalleryItem, ZhiGalleryProps } from "./types";
