+++ src/components/zhi/moment-detail/index.ts
@@ -1,0 +1,4 @@
+// Moment detail component barrel export
+
+export { ZhiMomentDetail as default, ZhiMomentDetail } from "./moment-detail-main";
+export type { MomentComment, MomentDetailProps } from "./types";
