+++ src/components/zhi/stats-dashboard.tsx
@@ -1,904 +1,7 @@
-"use client";
-/* eslint-disable react-hooks/set-state-in-effect */
+// Re-export from stats-dashboard directory for backward compatibility
+// This file is being deprecated in favor of the modular stats-dashboard/ directory
 
-import React, { useState, useEffect, useMemo } from "react";
-import { usePathname } from "next/navigation";
-import Link from "next/link";
-import {
-  PieChart,
-  Pie,
-  Cell,
-  BarChart,
-  Bar,
-  ResponsiveContainer,
-  Tooltip,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-} from "recharts";
-import {
-  Camera,
-  Activity,
-  Clock,
-  Film,
-  Gamepad2,
-  Code,
-  GitCommit,
-  ArrowUpRight,
-  Zap,
-} from "lucide-react";
-import { getLocaleFromPathname } from "@/lib/i18n";
-import type { DashboardStatsData } from "@/lib/dashboard-stats";
-import { NowPlayingCard, type MusicTrack } from "./now-playing-card";
-
-// Animated Counter Component
-function AnimatedCounter({
-  end,
-  duration = 2000,
-  trigger,
-}: {
-  end: number;
-  duration?: number;
-  trigger: number;
-}) {
-  const [count, setCount] = useState(0);
-
-  useEffect(() => {
-    let startTime: number | null = null;
-    let frameId: number;
-
-    const animate = (currentTime: number) => {
-      if (!startTime) startTime = currentTime;
-      const progress = Math.min((currentTime - startTime) / duration, 1);
-
-      if (progress < 1) {
-        setCount(Math.floor(Math.random() * end * 1.5));
-      } else {
-        setCount(end);
-      }
-
-      if (progress < 1) {
-        frameId = requestAnimationFrame(animate);
-      }
-    };
-
-    setCount(0);
-    frameId = requestAnimationFrame(animate);
-    return () => cancelAnimationFrame(frameId);
-  }, [end, duration, trigger]);
-
-  return <span>{count.toLocaleString()}</span>;
-}
-
-// Code Frequency Heatmap Component with Modern Design
-interface HeatmapDay {
-  date: string;
-  count: number;
-  level: number;
-  label: string;
-}
-
-interface CodeFrequencyHeatmapProps {
-  heatmapData: HeatmapDay[];
-  gitHubStats?: {
-    currentStreak: number;
-    commitsWeek: number;
-  };
-  locale: string;
-}
-
-function CodeFrequencyHeatmap({ heatmapData, gitHubStats, locale }: CodeFrequencyHeatmapProps) {
-  const [hoveredDay, setHoveredDay] = useState<{ day: HeatmapDay; x: number; y: number } | null>(null);
-
-  // Get color class based on level - purple-cyan gradient
-  const getLevelColor = (level: number) => {
-    switch (level) {
-      case 0: return 'bg-stone-100 dark:bg-stone-800/50';
-      case 1: return 'bg-violet-200 dark:bg-violet-900/50';
-      case 2: return 'bg-violet-400 dark:bg-violet-700/70';
-      case 3: return 'bg-fuchsia-500 dark:bg-fuchsia-600/80';
-      case 4: return 'bg-cyan-400 dark:bg-cyan-500';
-      default: return 'bg-stone-100 dark:bg-stone-800/50';
-    }
-  };
-
-  // Group data by weeks (7 days per week, starting from Sunday)
-  const weeks = useMemo(() => {
-    const result: HeatmapDay[][] = [];
-    for (let i = 0; i < heatmapData.length; i += 7) {
-      result.push(heatmapData.slice(i, i + 7));
-    }
-    return result;
-  }, [heatmapData]);
-
-  // Get month labels with positions
-  const monthLabels = useMemo(() => {
-    const labels: { label: string; weekIndex: number }[] = [];
-    let lastMonth = -1;
-
-    weeks.forEach((week, weekIndex) => {
-      // Check the first day of each week
-      const firstDay = week[0];
-      if (firstDay) {
-        const date = new Date(firstDay.date);
-        const month = date.getMonth();
-        if (month !== lastMonth) {
-          const monthName = date.toLocaleDateString(locale === 'zh' ? 'zh-CN' : 'en-US', { month: 'short' });
-          labels.push({ label: monthName, weekIndex });
-          lastMonth = month;
-        }
-      }
-    });
-
-    return labels;
-  }, [weeks, locale]);
-
-  // Calculate total contributions
-  const totalContributions = useMemo(() => {
-    return heatmapData.reduce((sum, day) => sum + day.count, 0);
-  }, [heatmapData]);
-
-  const weekDays = locale === 'zh' ? ['', '一', '', '三', '', '五', ''] : ['', 'Mon', '', 'Wed', '', 'Fri', ''];
-
-  const handleMouseEnter = (day: HeatmapDay, event: React.MouseEvent) => {
-    const rect = event.currentTarget.getBoundingClientRect();
-    setHoveredDay({
-      day,
-      x: rect.left + rect.width / 2,
-      y: rect.top - 8
-    });
-  };
-
-  return (
-    <div className="group col-span-1 min-h-[280px] overflow-hidden rounded-2xl border border-stone-100 bg-white p-4 shadow-sm transition-all duration-300 hover:shadow-md relative sm:p-5 dark:border-stone-800 dark:bg-stone-900">
-      {/* Subtle gradient background */}
-      <div
-        className="pointer-events-none absolute inset-0 opacity-[0.03] dark:opacity-[0.08]"
-        style={{
-          background: 'radial-gradient(ellipse at top right, #c084fc, transparent 50%), radial-gradient(ellipse at bottom left, #22d3ee, transparent 50%)'
-        }}
-      />
-
-      <div className="relative z-10 flex h-full flex-col">
-        {/* Header - Fixed layout to prevent wrapping */}
-        <div className="flex items-start justify-between mb-4 gap-2">
-          <div className="flex items-center gap-2 min-w-0">
-            <div className="rounded-xl bg-gradient-to-br from-violet-100 to-cyan-100 p-2 text-violet-600 flex-shrink-0 dark:from-violet-900/30 dark:to-cyan-900/30 dark:text-violet-400">
-              <GitCommit size={18} />
-            </div>
-            <div className="min-w-0">
-              <h4 className="font-serif text-base text-stone-800 whitespace-nowrap dark:text-stone-100">
-                {locale === 'zh' ? '代码频率' : 'Code Frequency'}
-              </h4>
-              <p className="flex items-center gap-1 text-[9px] font-bold uppercase tracking-wider text-stone-400 whitespace-nowrap">
-                <span className="relative flex h-1.5 w-1.5 flex-shrink-0">
-                  <span className="absolute inline-flex h-full w-full animate-ping rounded-full bg-cyan-400 opacity-75" />
-                  <span className="relative inline-flex h-1.5 w-1.5 rounded-full bg-cyan-500" />
-                </span>
-                {totalContributions.toLocaleString()} {locale === 'zh' ? '次提交' : 'commits'}
-              </p>
-            </div>
-          </div>
-
-          {gitHubStats && (
-            <div className="flex gap-3 flex-shrink-0">
-              <div className="text-right">
-                <div className="flex items-center justify-end gap-0.5 text-stone-400">
-                  <Zap size={10} className="text-amber-500" />
-                  <span className="text-[9px] whitespace-nowrap">{locale === 'zh' ? '连续' : 'Streak'}</span>
-                </div>
-                <span className="font-mono text-lg font-bold bg-gradient-to-r from-violet-600 to-cyan-500 bg-clip-text text-transparent">
-                  {gitHubStats.currentStreak}<span className="text-xs text-stone-400">d</span>
-                </span>
-              </div>
-              <div className="text-right">
-                <span className="block text-[9px] text-stone-400 whitespace-nowrap">{locale === 'zh' ? '本周' : 'Week'}</span>
-                <span className="font-mono text-xl font-bold text-cyan-500">
-                  {gitHubStats.commitsWeek}
-                </span>
-              </div>
-            </div>
-          )}
-        </div>
-
-        {/* Heatmap Container */}
-        <div className="flex-1 flex flex-col">
-          {/* Month Labels - Positioned above the grid with correct alignment */}
-          <div className="flex mb-1">
-            <div className="w-6 flex-shrink-0" /> {/* Spacer for week day labels */}
-            <div className="relative flex-1" style={{ height: '14px' }}>
-              {monthLabels.map(({ label, weekIndex }, idx) => (
-                <span
-                  key={idx}
-                  className="absolute text-[9px] font-medium text-stone-400 dark:text-stone-500"
-                  style={{ left: `${weekIndex * 15}px` }}
-                >
-                  {label}
-                </span>
-              ))}
-            </div>
-          </div>
-
-          {/* Heatmap Grid with Week Labels */}
-          <div className="flex overflow-x-auto pb-2 scrollbar-hide">
-            {/* Week day labels */}
-            <div className="flex flex-col gap-[3px] pr-1 flex-shrink-0">
-              {weekDays.map((day, idx) => (
-                <div key={idx} className="h-[12px] flex items-center justify-end">
-                  <span className="text-[8px] font-medium text-stone-400 dark:text-stone-500 w-5 text-right">
-                    {day}
-                  </span>
-                </div>
-              ))}
-            </div>
-
-            {/* Grid - Circular dots style */}
-            <div className="flex gap-[3px]">
-              {weeks.map((week, weekIdx) => (
-                <div key={weekIdx} className="flex flex-col gap-[3px]">
-                  {week.map((day, dayIdx) => (
-                    <div
-                      key={dayIdx}
-                      className={`h-[12px] w-[12px] rounded-full transition-all cursor-pointer
-                        hover:ring-2 hover:ring-violet-400/50 hover:scale-125
-                        ${getLevelColor(day.level)}`}
-                      onMouseEnter={(e) => handleMouseEnter(day, e)}
-                      onMouseLeave={() => setHoveredDay(null)}
-                    />
-                  ))}
-                  {/* Fill empty days at end of last week */}
-                  {week.length < 7 && Array.from({ length: 7 - week.length }).map((_, idx) => (
-                    <div key={`empty-${idx}`} className="h-[12px] w-[12px]" />
-                  ))}
-                </div>
-              ))}
-            </div>
-          </div>
-
-          {/* Legend */}
-          <div className="flex items-center justify-between mt-3 pt-2 border-t border-stone-100 dark:border-stone-800">
-            <div className="flex items-center gap-1.5">
-              <span className="text-[9px] text-stone-400">{locale === 'zh' ? '少' : 'Less'}</span>
-              <div className="flex gap-[2px]">
-                {[0, 1, 2, 3, 4].map((level) => (
-                  <div key={level} className={`h-[8px] w-[8px] rounded-full ${getLevelColor(level)}`} />
-                ))}
-              </div>
-              <span className="text-[9px] text-stone-400">{locale === 'zh' ? '多' : 'More'}</span>
-            </div>
-            <a
-              href="https://github.com/ZhiHao-He"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="flex items-center gap-0.5 text-[9px] text-stone-400 hover:text-violet-500 transition-colors"
-            >
-              <span>@GitHub</span>
-              <ArrowUpRight size={9} />
-            </a>
-          </div>
-        </div>
-      </div>
-
-      {/* Floating Tooltip */}
-      {hoveredDay && (
-        <div
-          className="fixed z-50 pointer-events-none animate-in fade-in zoom-in-95 duration-150"
-          style={{
-            left: hoveredDay.x,
-            top: hoveredDay.y,
-            transform: 'translate(-50%, -100%)'
-          }}
-        >
-          <div className="bg-stone-900 dark:bg-stone-100 text-white dark:text-stone-900 px-3 py-2 rounded-lg shadow-xl text-center">
-            <div className="font-bold text-sm">
-              {hoveredDay.day.count} {locale === 'zh' ? '次提交' : hoveredDay.day.count === 1 ? 'contribution' : 'contributions'}
-            </div>
-            <div className="text-[10px] text-stone-400 dark:text-stone-500 mt-0.5">
-              {new Date(hoveredDay.day.date).toLocaleDateString(locale === 'zh' ? 'zh-CN' : 'en-US', {
-                weekday: 'short',
-                year: 'numeric',
-                month: 'short',
-                day: 'numeric'
-              })}
-            </div>
-            {/* Tooltip Arrow */}
-            <div className="absolute left-1/2 -translate-x-1/2 top-full w-0 h-0 border-l-[6px] border-r-[6px] border-t-[6px] border-l-transparent border-r-transparent border-t-stone-900 dark:border-t-stone-100" />
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Types
-interface StatCardData {
-  icon: React.ReactNode;
-  title: string;
-  subtitle: string;
-  value: string | number;
-  trend?: "up" | "down" | "stable";
-  href?: string;
-  color: string;
-}
-
-interface ZhiStatsDashboardProps {
-  stats: DashboardStatsData;
-  highlights?: StatCardData[];
-}
-
-
-export function ZhiStatsDashboard({
-  stats,
-  highlights,
-}: ZhiStatsDashboardProps) {
-  const pathname = usePathname();
-  const locale = getLocaleFromPathname(pathname) ?? "en";
-  const [animationTrigger, setAnimationTrigger] = useState(0);
-  const [isFlashing, setIsFlashing] = useState(false);
-
-  const t = (key: string) => {
-    const translations: Record<string, Record<string, string>> = {
-      en: {
-        "Life Log": "Life Log",
-        subtitle: "Quantifying the hobbies that keep me sane.",
-        "Shutter Count": "Shutter Count",
-        "The Balance": "The Balance",
-        "Weekly Routine": "Weekly Routine",
-        Movies: "Movies",
-        Games: "Games",
-        "Now Playing": "Now Playing",
-        Progress: "Progress",
-        "The Journey": "The Journey",
-        "Daily Steps": "Daily Steps",
-        "Avg. Steps": "Avg. Steps",
-        "The Output": "The Output",
-        "Current Focus Areas": "Current Focus Areas",
-        "View Details": "View Details",
-      },
-      zh: {
-        "Life Log": "生活记录",
-        subtitle: "量化那些让我保持理智的爱好。",
-        "Shutter Count": "快门次数",
-        "The Balance": "平衡",
-        "Weekly Routine": "每周日程",
-        Movies: "电影",
-        Games: "游戏",
-        "Now Playing": "正在玩",
-        Progress: "进度",
-        "The Journey": "旅程",
-        "Daily Steps": "每日步数",
-        "Avg. Steps": "平均步数",
-        "The Output": "产出",
-        "Current Focus Areas": "当前关注领域",
-        "View Details": "查看详情",
-      },
-    };
-    return translations[locale]?.[key] || key;
-  };
-
-  const handleShutterClick = () => {
-    setIsFlashing(true);
-    setTimeout(() => {
-      setAnimationTrigger((prev) => prev + 1);
-    }, 100);
-    setTimeout(() => {
-      setIsFlashing(false);
-    }, 300);
-  };
-
-  const avgSteps = Math.round(
-    stats.stepsData.entries.reduce((acc: number, curr: { steps: number }) => acc + curr.steps, 0) / (stats.stepsData.entries.length || 1)
-  );
-
-  // Aggregate daily contributions for Standard Heatmap (Last 3 Months / ~90 Days)
-  const heatmapData = useMemo(() => {
-    if (!stats.gitHubContributions) return [];
-
-    // Create map for O(1) lookup
-    const dataMap = new Map((stats.gitHubContributions || []).map(i => {
-      // Ensure date string format YYYY-MM-DD
-      const dateStr = typeof i.date === 'string' ? i.date.split('T')[0] : new Date(i.date).toISOString().split('T')[0];
-      return [dateStr, i.value];
-    }));
-
-    const today = new Date();
-    // Start from 3 months ago (~90 days)
-    const startDate = new Date(today);
-    startDate.setDate(today.getDate() - 90);
-
-    // Align start date to the previous Sunday (to start grid correctly)
-    const dayOfWeek = startDate.getDay(); // 0 = Sunday
-    const alignedStartDate = new Date(startDate);
-    alignedStartDate.setDate(startDate.getDate() - dayOfWeek);
-
-    const days = [];
-    // 14 weeks * 7 days = 98 grid cells (covers ~3 months with alignment)
-    const totalDays = 14 * 7;
-
-    for (let i = 0; i < totalDays; i++) {
-      const d = new Date(alignedStartDate);
-      d.setDate(alignedStartDate.getDate() + i);
-      // Assert dateStr is string to avoid lint error
-      const dateStr = d.toISOString().split('T')[0] as string;
-
-      const isFuture = d > new Date();
-      const count = isFuture ? 0 : (dataMap.get(dateStr) || 0);
-
-      let level = 0;
-      if (count > 0) level = 1;
-      if (count >= 3) level = 2;
-      if (count >= 6) level = 3;
-      if (count >= 10) level = 4;
-
-      days.push({
-        date: dateStr,
-        count,
-        level,
-        label: new Date(dateStr).toLocaleDateString(undefined, { weekday: 'short', month: 'short', day: 'numeric' })
-      });
-    }
-    return days;
-  }, [stats.gitHubContributions]);
-
-  return (
-    <div className="w-full animate-in fade-in pb-16 duration-700">
-      {/* CSS Animations */}
-      <style>{`
-        @keyframes flash {
-          0% { opacity: 0; }
-          10% { opacity: 1; }
-          100% { opacity: 0; }
-        }
-        .flash-overlay {
-          animation: flash 0.3s ease-out;
-          pointer-events: none;
-        }
-      `}</style>
-
-      <div className="mx-auto grid max-w-5xl grid-cols-1 gap-3 px-4 pt-6 sm:grid-cols-2 sm:gap-4 lg:grid-cols-3 lg:gap-6">
-        {/* Card 1: Interactive Shutter Count - Always spans full width */}
-        <div className="relative col-span-1 sm:col-span-2 select-none overflow-hidden rounded-2xl border border-stone-800 bg-[#171717] p-4 text-white shadow-2xl sm:p-6 lg:col-span-2">
-          {/* Flash Overlay */}
-          {isFlashing && (
-            <div className="flash-overlay absolute inset-0 z-50 bg-white"></div>
-          )}
-
-          {/* Background Texture */}
-          <div
-            className="pointer-events-none absolute inset-0 opacity-20"
-            style={{
-              backgroundImage: "radial-gradient(#333 1px, transparent 0)",
-              backgroundSize: "10px 10px",
-            }}
-          ></div>
-
-          {/* HUD Overlay */}
-          <div className="absolute left-4 top-4 h-4 w-4 border-l-2 border-t-2 border-cyan-500/50"></div>
-          <div className="absolute right-4 top-4 h-4 w-4 border-r-2 border-t-2 border-cyan-500/50"></div>
-          <div className="absolute bottom-4 left-4 h-4 w-4 border-b-2 border-l-2 border-cyan-500/50"></div>
-          <div className="absolute bottom-4 right-4 h-4 w-4 border-b-2 border-r-2 border-cyan-500/50"></div>
-
-          {/* AF Focus Box Center */}
-          <div
-            className={`absolute left-1/2 top-1/2 h-12 w-16 -translate-x-1/2 -translate-y-1/2 border transition-all duration-100 ${isFlashing ? "h-10 w-14 scale-90 border-green-500" : "scale-100 border-yellow-500/50"}`}
-          >
-            <div className="absolute left-1/2 top-1/2 h-1 w-1 -translate-x-1/2 -translate-y-1/2 bg-yellow-500/50"></div>
-          </div>
-
-          {/* Header Row */}
-          <div className="relative z-20 mb-6 flex items-center justify-between">
-            <div className="flex items-center gap-4">
-              <div
-                className={`flex items-center gap-2 font-mono text-xs font-bold transition-colors ${isFlashing ? "text-red-500" : "text-stone-500"}`}
-              >
-                <div
-                  className={`h-2 w-2 rounded-full ${isFlashing ? "bg-red-500" : "bg-stone-500"}`}
-                ></div>
-                REC
-              </div>
-              <div className="rounded border border-stone-700 bg-stone-900 px-2 py-0.5 font-mono text-[10px] text-stone-400">
-                1/8000s
-              </div>
-            </div>
-
-            {/* Shutter Button */}
-            <button
-              onClick={handleShutterClick}
-              className="group/btn relative flex items-center justify-center"
-              title="Capture Frame"
-            >
-              <div className="flex h-12 w-12 items-center justify-center rounded-full border-2 border-stone-600 bg-gradient-to-b from-stone-700 to-stone-800 shadow-lg transition-transform active:scale-95">
-                <div className="h-8 w-8 rounded-full border border-stone-800 bg-stone-900"></div>
-              </div>
-            </button>
-          </div>
-
-          {/* Main Content Area */}
-          <div className="relative z-10 flex min-h-64 w-full flex-col gap-6 md:flex-row">
-            {/* Left: Stats */}
-            <div className="flex w-full flex-col justify-end border-r border-stone-800/50 pb-4 pr-4 md:w-1/3">
-              <div className="mb-1 flex items-center gap-2 text-cyan-500/80">
-                <Camera size={16} />
-                <span className="text-[10px] font-bold uppercase tracking-widest">
-                  {t("Shutter Count")}
-                </span>
-              </div>
-              <div className="font-mono text-6xl font-bold leading-none tracking-tighter text-white tabular-nums">
-                <AnimatedCounter
-                  end={stats.photoCount}
-                  trigger={animationTrigger}
-                  duration={1000}
-                />
-              </div>
-              <div className="mt-4 space-y-1 font-mono text-[10px] text-stone-500">
-                <div className="flex justify-between">
-                  <span>BUFFER</span>
-                  <span className="text-cyan-500">{isFlashing ? "BUSY" : "READY"}</span>
-                </div>
-              </div>
-            </div>
-
-            {/* Right: Chart */}
-            <div className="relative flex-1">
-              <ResponsiveContainer width="100%" height="100%" minWidth={1} minHeight={1}>
-                <BarChart
-                  key={animationTrigger}
-                  data={stats.photosByWeek}
-                  margin={{ top: 20, right: 10, left: -20, bottom: 0 }}
-                  barCategoryGap="20%"
-                >
-                  <CartesianGrid strokeDasharray="3 3" stroke="#222" vertical={false} />
-                  <XAxis
-                    dataKey="day"
-                    axisLine={false}
-                    tickLine={false}
-                    tick={{ fill: "#555", fontSize: 10, fontFamily: "monospace" }}
-                    dy={10}
-                  />
-                  <YAxis
-                    axisLine={false}
-                    tickLine={false}
-                    tick={{ fill: "#333", fontSize: 9, fontFamily: "monospace" }}
-                  />
-                  <Tooltip
-                    cursor={{ fill: "rgba(255, 255, 255, 0.05)" }}
-                    contentStyle={{
-                      background: "rgba(23, 23, 23, 0.9)",
-                      border: "1px solid #333",
-                      borderRadius: "8px",
-                      fontFamily: "monospace",
-                      fontSize: "12px",
-                    }}
-                  />
-                  <Bar dataKey="count" fill="#22d3ee" radius={[4, 4, 0, 0]} />
-                </BarChart>
-              </ResponsiveContainer>
-            </div>
-          </div>
-        </div>
-
-        {/* Card 2: Weekly Routine */}
-        <div className="group col-span-1 rounded-2xl border border-stone-100 bg-white p-4 shadow-sm transition-all duration-300 hover:shadow-md sm:p-6 dark:border-stone-800 dark:bg-stone-900">
-          <div className="mb-4 flex items-center gap-3">
-            <div className="rounded-xl bg-indigo-50 p-2 text-indigo-500 transition-colors group-hover:bg-indigo-100 dark:bg-indigo-900/10 dark:group-hover:bg-indigo-900/20">
-              <Clock size={20} strokeWidth={1.5} />
-            </div>
-            <div>
-              <h4 className="font-serif text-lg text-stone-800 dark:text-stone-100">
-                {t("The Balance")}
-              </h4>
-              <p className="text-[10px] font-bold uppercase tracking-wider text-stone-400">
-                {t("Weekly Routine")}
-              </p>
-            </div>
-          </div>
-          <div className="relative h-40 min-h-40 w-full">
-            <ResponsiveContainer width="100%" height="100%" minWidth={1} minHeight={1}>
-              <PieChart>
-                <Pie
-                  data={stats.routineData}
-                  cx="50%"
-                  cy="50%"
-                  innerRadius={50}
-                  outerRadius={70}
-                  paddingAngle={5}
-                  dataKey="value"
-                  stroke="none"
-                >
-                  {stats.routineData.map((entry, index) => (
-                    <Cell key={`cell-${index}`} fill={entry.color} />
-                  ))}
-                </Pie>
-                <Tooltip
-                  contentStyle={{
-                    borderRadius: "8px",
-                    border: "none",
-                    boxShadow: "0 4px 12px rgba(0,0,0,0.1)",
-                  }}
-                />
-              </PieChart>
-            </ResponsiveContainer>
-            <div className="pointer-events-none absolute inset-0 flex items-center justify-center">
-              <span className="font-serif text-sm italic text-stone-400">24h</span>
-            </div>
-          </div>
-        </div>
-
-        {/* Card 2.5: Now Playing Music - Paired with Weekly Routine on mobile */}
-        <div className="col-span-1">
-          <NowPlayingCard
-            track={stats.nowPlaying?.[0] as MusicTrack | null}
-            recentTracks={(stats.nowPlaying?.slice(1) ?? []) as MusicTrack[]}
-            locale={locale}
-          />
-        </div>
-
-        {/* Card 3: Media Diet - Spans full width for complex charts */}
-        <div className="col-span-1 sm:col-span-2 flex flex-col overflow-hidden rounded-2xl border border-stone-100 bg-white shadow-sm md:flex-row lg:col-span-2 dark:border-stone-800 dark:bg-stone-900">
-          {/* Movies Section */}
-          <div className="group flex-1 border-b border-stone-100 p-6 transition-colors hover:bg-stone-50 md:border-b-0 md:border-r dark:border-stone-800 dark:hover:bg-stone-800/30">
-            <div className="mb-4 flex items-center justify-between">
-              <div className="flex items-center gap-3">
-                <div className="rounded-xl bg-rose-50 p-2 text-rose-500 dark:bg-rose-900/10">
-                  <Film size={18} />
-                </div>
-                <h4 className="font-serif text-md text-stone-800 dark:text-stone-100">
-                  {t("Movies")}
-                </h4>
-              </div>
-              <span className="text-2xl font-bold text-stone-800 dark:text-stone-200">
-                {stats.movieCount}
-              </span>
-            </div>
-            <div className="h-24 min-h-24">
-              <ResponsiveContainer width="100%" height="100%" minWidth={1} minHeight={1}>
-                <BarChart data={stats.movieData}>
-                  <Bar dataKey="movies" fill="#fb7185" radius={[2, 2, 0, 0]} />
-                  <XAxis
-                    dataKey="month"
-                    fontSize={10}
-                    tickLine={false}
-                    axisLine={false}
-                    tick={{ fill: "#a8a29e" }}
-                  />
-                  <Tooltip
-                    cursor={{ fill: "transparent" }}
-                    contentStyle={{
-                      borderRadius: "8px",
-                      border: "none",
-                      boxShadow: "0 4px 12px rgba(0,0,0,0.05)",
-                    }}
-                  />
-                </BarChart>
-              </ResponsiveContainer>
-            </div>
-          </div>
-
-          {/* Games Section */}
-          <div className="group flex-1 p-6 transition-colors hover:bg-stone-50 dark:hover:bg-stone-800/30">
-            <div className="mb-4 flex items-center justify-between">
-              <div className="flex items-center gap-3">
-                <div className="rounded-xl bg-emerald-50 p-2 text-emerald-500 dark:bg-emerald-900/10">
-                  <Gamepad2 size={18} />
-                </div>
-                <h4 className="font-serif text-md text-stone-800 dark:text-stone-100">
-                  {t("Games")}
-                </h4>
-              </div>
-              {stats.currentGame && (
-                <div className="text-right">
-                  <span className="block text-xs text-stone-400">{t("Now Playing")}</span>
-                  <span className="text-sm font-bold text-stone-800 dark:text-stone-200">
-                    {stats.currentGame.name}
-                  </span>
-                </div>
-              )}
-            </div>
-            {stats.currentGame && (
-              <>
-                <div className="mt-2">
-                  <div className="mb-1 flex justify-between text-xs text-stone-500">
-                    <span>{t("Progress")}</span>
-                    <span>{stats.currentGame.progress}%</span>
-                  </div>
-                  <div className="h-2 w-full overflow-hidden rounded-full bg-stone-100 dark:bg-stone-800">
-                    <div
-                      className="h-full rounded-full bg-emerald-500"
-                      style={{ width: `${stats.currentGame.progress}%` }}
-                    ></div>
-                  </div>
-                </div>
-                <div className="mt-4 flex gap-2">
-                  {["RPG", "Strategy"].map((tag) => (
-                    <span
-                      key={tag}
-                      className="rounded border border-stone-200 bg-stone-100 px-2 py-1 text-[10px] text-stone-500 dark:border-stone-700 dark:bg-stone-800"
-                    >
-                      {tag}
-                    </span>
-                  ))}
-                </div>
-              </>
-            )}
-          </div>
-        </div>
-
-        {/* Card 4: Daily Steps */}
-        <div className="group col-span-1 rounded-2xl border border-stone-100 bg-white p-4 shadow-sm transition-all duration-300 hover:shadow-md sm:p-6 dark:border-stone-800 dark:bg-stone-900">
-          <div className="mb-4 flex items-center justify-between">
-            <div className="flex items-center gap-3">
-              <div className="rounded-xl bg-sage-50 p-2 text-sage-600 transition-colors group-hover:bg-sage-100 dark:bg-sage-900/10 dark:text-sage-400 dark:group-hover:bg-sage-900/20">
-                <Activity size={20} strokeWidth={1.5} />
-              </div>
-              <div>
-                <h4 className="font-serif text-lg text-stone-800 dark:text-stone-100">
-                  {t("The Journey")}
-                </h4>
-                <p className="text-[10px] font-bold uppercase tracking-wider text-stone-400">
-                  {t("Daily Steps")}
-                </p>
-              </div>
-            </div>
-            <div className="text-right">
-              <span className="block text-xl font-bold text-stone-800 dark:text-stone-100">
-                {avgSteps.toLocaleString()}
-              </span>
-              {stats.stepsData.startDate && stats.stepsData.endDate && (
-                <span className="block text-[10px] text-sage-600 dark:text-sage-400">
-                  {stats.stepsData.startDate} - {stats.stepsData.endDate}
-                </span>
-              )}
-            </div>
-          </div>
-          <div className="h-24 min-h-24 w-full">
-            <ResponsiveContainer width="100%" height="100%" minWidth={1} minHeight={1}>
-              <BarChart data={stats.stepsData.entries}>
-                <Tooltip
-                  cursor={{ fill: "transparent" }}
-                  contentStyle={{
-                    borderRadius: "8px",
-                    border: "none",
-                    boxShadow: "0 4px 12px rgba(0,0,0,0.05)",
-                    backgroundColor: "#fff",
-                    color: "#333",
-                    padding: "8px 12px",
-                  }}
-                  content={({ active, payload }) => {
-                    if (active && payload && payload.length > 0) {
-                      const data = payload[0]?.payload as { dayNum: number; steps: number } | undefined;
-                      if (!data) return null;
-                      return (
-                        <div className="rounded-lg bg-white px-3 py-2 shadow-lg dark:bg-stone-800">
-                          <p className="text-sm font-medium text-stone-600 dark:text-stone-300">
-                            {data.dayNum}
-                          </p>
-                          <p className="text-lg font-bold text-sage-600">
-                            {data.steps.toLocaleString()}
-                          </p>
-                        </div>
-                      );
-                    }
-                    return null;
-                  }}
-                />
-                <Bar dataKey="steps" fill="#5c9c6d" radius={[4, 4, 4, 4]} barSize={8} />
-              </BarChart>
-            </ResponsiveContainer>
-          </div>
-        </div>
-
-        {/* Card 5: Languages (Donut) - Paired with Daily Steps on mobile */}
-        <div className="group col-span-1 rounded-2xl border border-stone-100 bg-white p-4 shadow-sm transition-all duration-300 hover:shadow-md sm:p-6 dark:border-stone-800 dark:bg-stone-900">
-          <div className="mb-4 flex items-center gap-3">
-            <div className="rounded-xl bg-orange-50 p-2 text-orange-500 transition-colors group-hover:bg-orange-100 dark:bg-orange-900/10 dark:text-orange-400 dark:group-hover:bg-orange-900/20">
-              <Code size={20} strokeWidth={1.5} />
-            </div>
-            <div>
-              <h4 className="font-serif text-lg text-stone-800 dark:text-stone-100">
-                {t("The Output")}
-              </h4>
-              <p className="text-[10px] font-bold uppercase tracking-wider text-stone-400">
-                Languages
-              </p>
-            </div>
-          </div>
-          <div className="relative h-40 min-h-40 w-full">
-            <ResponsiveContainer width="100%" height="100%" minWidth={1} minHeight={1}>
-              <PieChart>
-                <Pie
-                  data={stats.skillData.map((skill) => {
-                    const colors: Record<string, string> = {
-                      "Python": "#3572A5",
-                      "TypeScript": "#3178c6",
-                      "Jupyter Notebook": "#DA5B0B",
-                      "HTML": "#e34c26",
-                      "CSS": "#563d7c",
-                      "JavaScript": "#f1e05a"
-                    };
-                    return { ...skill, color: colors[skill.name] || "#a8a29e" };
-                  })}
-                  cx="50%"
-                  cy="50%"
-                  innerRadius={50}
-                  outerRadius={70}
-                  paddingAngle={5}
-                  dataKey="level"
-                  stroke="none"
-                >
-                  {stats.skillData.map((entry, index) => {
-                    const colors: Record<string, string> = {
-                      "Python": "#3572A5",
-                      "TypeScript": "#3178c6",
-                      "Jupyter Notebook": "#DA5B0B",
-                      "HTML": "#e34c26",
-                      "CSS": "#563d7c",
-                      "JavaScript": "#f1e05a"
-                    };
-                    const color = colors[entry.name] || "#a8a29e";
-                    return <Cell key={`cell-${index}`} fill={color} />;
-                  })}
-                </Pie>
-                <Tooltip
-                  contentStyle={{
-                    borderRadius: "8px",
-                    border: "none",
-                    boxShadow: "0 4px 12px rgba(0,0,0,0.1)",
-                  }}
-                />
-              </PieChart>
-            </ResponsiveContainer>
-            <div className="pointer-events-none absolute inset-0 flex flex-col items-center justify-center">
-              <span className="text-3xl font-bold text-stone-800 dark:text-stone-100">
-                {stats.skillData.length}
-              </span>
-              <span className="text-[10px] font-medium uppercase tracking-wider text-stone-400">
-                Langs
-              </span>
-            </div>
-          </div>
-        </div>
-
-        {/* Card 6: Code Frequency (Enhanced Heatmap) - Full width */}
-        {stats.gitHubContributions && stats.gitHubContributions.length > 0 && (
-          <CodeFrequencyHeatmap
-            heatmapData={heatmapData}
-            gitHubStats={stats.gitHubStats}
-            locale={locale}
-          />
-        )}
-
-        {/* Quick Links */}
-        {highlights && highlights.length > 0 && (
-          <div className="col-span-1 sm:col-span-2 lg:col-span-3">
-            <h3 className="mb-4 font-serif text-xl text-stone-800 dark:text-stone-100">
-              {t("View Details")}
-            </h3>
-            <div className="grid grid-cols-2 gap-4 md:grid-cols-4">
-              {highlights.map((item, index) => (
-                <Link
-                  key={index}
-                  href={item.href || "#"}
-                  className="group flex items-center gap-3 rounded-xl border border-stone-100 bg-white p-4 transition-all hover:border-sage-200 hover:shadow-md dark:border-stone-800 dark:bg-stone-900 dark:hover:border-sage-700"
-                >
-                  <div className={`rounded-lg p-2 ${item.color}`}>{item.icon}</div>
-                  <div className="flex-1">
-                    <div className="text-sm font-medium text-stone-800 dark:text-stone-100">
-                      {item.title}
-                    </div>
-                    <div className="text-xs text-stone-500">{item.value}</div>
-                  </div>
-                  <ArrowUpRight
-                    size={16}
-                    className="text-stone-400 transition-transform group-hover:translate-x-0.5 group-hover:-translate-y-0.5"
-                  />
-                </Link>
-              ))}
-            </div>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-export default ZhiStatsDashboard;
+export { ZhiStatsDashboard as default, ZhiStatsDashboard } from "./stats-dashboard/stats-dashboard-main";
+export { AnimatedCounter } from "./stats-dashboard/animated-counter";
+export { CodeFrequencyHeatmap } from "./stats-dashboard/code-frequency-heatmap";
+export type { HeatmapDay, StatCardData } from "./stats-dashboard/types";
