+++ src/components/zhi/stats-dashboard/index.ts
@@ -1,0 +1,6 @@
+// Stats dashboard component barrel export
+
+export { ZhiStatsDashboard as default, ZhiStatsDashboard } from "./stats-dashboard-main";
+export { AnimatedCounter } from "./animated-counter";
+export { CodeFrequencyHeatmap } from "./code-frequency-heatmap";
+export type { HeatmapDay, CodeFrequencyHeatmapProps, StatCardData } from "./types";
